Initial changes for ERC
The event rate controller is new in gen 4 sensors. It throttles the
number of events coming out of the camera. The exact implementation is
unknown and not straight forward (it impacts the events below that event
rate). From testing it appears that this rate is the target upper limit.

- erc_mode - the mode of the controller - enabled, disabled, or na (for
if you have an old sensor)
- erc_rate - the rate at which the controller should throttle